+++ cygwin&babun/install_oh_myzsh.py
@@ -15,14 +15,17 @@ print n*'='
 print title
 print n*'='
 
-cmds = [
+if os.path.exists("/home/lingjie/.oh-my-zsh"):
+	print "oh-my-zsh has installed."
+else :
+	cmds = [
 		"git clone git://github.com/robbyrussell/oh-my-zsh.git ~/.oh-my-zsh",
-	    "cp ~/.oh-my-zsh/templates/zshrc.zsh-template ~/.zshrc"
-]
+		"cp ~/.oh-my-zsh/templates/zshrc.zsh-template ~/.zshrc"
+	]
 
-for cmd in cmds:
-	print cmd
-	os.system(cmd)
+	for cmd in cmds:
+		print cmd
+		os.system(cmd)
 
 print n*'='    
 print "= Done!" + (n-len("= Done!")-1)*' ' + "="
